refactor(ai): move chat title generation service
- Extract title prompt, retry, and fallback logic from the chat router

- Keep router handling focused on chat lookup, usage tracking, and persistence
